* Split experiment and analysis.
